Split payments handlers by workflow
