+++ apps/sdp-api/src/routes/payments/handlers/balances.ts
@@ -1,0 +1,100 @@
+import { formatDecimalAmount } from "@/lib/amount";
+import { AppError } from "@/lib/errors";
+import { success } from "@/lib/response";
+import { createRpc, getAccountInfo } from "@/services/solana/rpc";
+import type { Address } from "@solana/kit";
+import { type AppContext, getPaymentsRepository } from "../context";
+import {
+  DESTINATION_ALLOWLIST_POLICY_TYPE,
+  PAYMENT_POLICY_VERSION,
+  TRANSFER_LIMITS_POLICY_TYPE,
+  buildWalletPolicyPayload,
+} from "../policy";
+import { updateWalletPolicySchema } from "../schemas";
+import { SOL_MINT, getSplTokenBalances } from "../token-accounts";
+import { resolveWalletFromParams } from "./transfers";
+
+export async function getWalletBalances(c: AppContext) {
+  const { wallet } = await resolveWalletFromParams(c, ["wallets:read"]);
+
+  const rpc = createRpc(c.env);
+  const accountInfo = await getAccountInfo(rpc, wallet.publicKey as Address);
+  const splBalances = await getSplTokenBalances(rpc, wallet.publicKey as Address);
+  const lamports = accountInfo?.lamports ?? 0n;
+
+  return success(c, {
+    walletBalances: {
+      walletId: wallet.walletId,
+      address: wallet.publicKey,
+      balances: [
+        {
+          token: "SOL",
+          mint: SOL_MINT,
+          amount: lamports.toString(),
+          uiAmount: formatDecimalAmount(lamports, 9),
+          decimals: 9,
+        },
+        ...splBalances,
+      ],
+    },
+  });
+}
+
+export async function getWalletPolicy(c: AppContext) {
+  const { wallet } = await resolveWalletFromParams(c, ["wallets:read"]);
+  const repository = getPaymentsRepository(c);
+
+  const rows = await repository.getWalletPoliciesByCustodyWalletId(wallet.id);
+  const payload = buildWalletPolicyPayload(wallet.walletId, rows, wallet.createdAt);
+
+  return success(c, { policy: payload });
+}
+
+export async function updateWalletPolicy(c: AppContext) {
+  const { wallet } = await resolveWalletFromParams(c, ["wallets:write"]);
+  const repository = getPaymentsRepository(c);
+
+  const body = await c.req.json();
+  const parsed = updateWalletPolicySchema.safeParse(body);
+
+  if (!parsed.success) {
+    throw new AppError("BAD_REQUEST", "Invalid request body", {
+      errors: parsed.error.flatten().fieldErrors,
+    });
+  }
+
+  const now = new Date().toISOString();
+  const rows = await repository.upsertWalletPolicies([
+    {
+      id: `pwp_${crypto.randomUUID()}`,
+      custodyWalletId: wallet.id,
+      policyType: DESTINATION_ALLOWLIST_POLICY_TYPE,
+      policy: JSON.stringify({
+        version: PAYMENT_POLICY_VERSION,
+        destinationAllowlist: parsed.data.destinationAllowlist,
+      }),
+      createdAt: now,
+      updatedAt: now,
+    },
+    {
+      id: `pwp_${crypto.randomUUID()}`,
+      custodyWalletId: wallet.id,
+      policyType: TRANSFER_LIMITS_POLICY_TYPE,
+      policy: JSON.stringify({
+        version: PAYMENT_POLICY_VERSION,
+        maxTransferAmount: parsed.data.maxTransferAmount ?? null,
+        maxDailyAmount: parsed.data.maxDailyAmount ?? null,
+      }),
+      createdAt: now,
+      updatedAt: now,
+    },
+  ]);
+
+  if (rows.length === 0) {
+    throw new AppError("INTERNAL_ERROR", "Failed to persist wallet policy");
+  }
+
+  const payload = buildWalletPolicyPayload(wallet.walletId, rows, now);
+
+  return success(c, { policy: payload });
+}
